Unit-Tests-And-Coverage-Report-Added
The unit tests were generated with CodePro and refined to improve code
coverage for the CoffeeMaker.java, Recipe.java, and Inventory.java
classes.

The code coverage report was generated with Eclipse Emma. Note that in
the statement coverage report, the relevant numbers for line coverage
can be found in the CoffeeMakerTest project, not in the CoffeeMaker
project!